Task Submission

+++ restaurants/views.py
@@ -3,6 +3,7 @@ from .models import Restaurant, Item
 from .forms import RestaurantForm, ItemForm, SignupForm, SigninForm
 from django.contrib.auth import login, authenticate, logout
 
+
 def signup(request):
     form = SignupForm()
     if request.method == 'POST':
@@ -59,6 +60,8 @@ def restaurant_detail(request, restaurant_id):
     return render(request, 'detail.html', context)
 
 def restaurant_create(request):
+    if request.user.is_anonymous:
+        return redirect('signin')
     form = RestaurantForm()
     if request.method == "POST":
         form = RestaurantForm(request.POST, request.FILES)
@@ -90,6 +93,10 @@ def item_create(request, restaurant_id):
 
 def restaurant_update(request, restaurant_id):
     restaurant_obj = Restaurant.objects.get(id=restaurant_id)
+
+    if not (request.user.is_staff or request.user == restaurant_obj.owner):
+        return redirect('no access')
+
     form = RestaurantForm(instance=restaurant_obj)
     if request.method == "POST":
         form = RestaurantForm(request.POST, request.FILES, instance=restaurant_obj)
@@ -103,6 +110,10 @@ def restaurant_update(request, restaurant_id):
     return render(request, 'update.html', context)
 
 def restaurant_delete(request, restaurant_id):
+
     restaurant_obj = Restaurant.objects.get(id=restaurant_id)
+    if not (request.user.is_staff):
+        return redirect('no access')
+        
     restaurant_obj.delete()
     return redirect('restaurant-list')
